Update Timelog.md
We finalized our doc and realized a flaw in our design. Alan and Talha talked that through for a long time and eventually came to a good solution. Yinuo and Vardan were also present for discussions and contributed. Denis, Vardan, Talha, Yinuo, and Alan all edited the document for general improvements and to match the latest specifications per the aforementioned discussion.